# Optimize ledstip by disabling unassigned overlays