--- a/src/main/io/ledstrip.h
+++ b/src/main/io/ledstrip.h
@@ -187,3 +187,5 @@ void applyDefaultLedStripConfig(ledConfig_t *ledConfig);
 void applyDefaultColors(hsvColor_t *colors);
 void applyDefaultModeColors(modeColorIndexes_t *modeColors);
 void applyDefaultSpecialColors(specialColorIndexes_t *specialColors);
+
+void updateRequiredOverlay(void);