Needs Review.

The current design continuously runs timer loops for all overlays regardless of whether or not any LEDs are actually using the overlay.  Then each overlay function iterates through all configured LED's to apply if needed.

This optimization "looks ahead" and turns off any overlays that are not assigned to any LEDs.  The most optimization is achieved when no overlays are used, and regresses to match the current performance level if every overlay is assigned.

Performance statistics gathered from the average ledstrip task time captured from `tasks`:

**F4** (REVO)

test | BF3.3RC2 | Optimized
---- |-------- | ---------
No LED's assigned | 105us | 7us
4 LED's (solid color) | 109us | 9us
32 LED's (solid color) | 132us | 12us

**F3** (SPRACINGF3)

test | BF3.3RC2 | Optimized
---- |-------- | ---------
No LED's assigned | 260us | 22us
4 LED's (solid color) | 272us | 26us
32 LED's (solid color) | 347us | 30us

I was able to test all of the overlays to confirm they're working correctly except for VTX as I don't have any appropriate hardware.