@mikeller 2018-02-22
Fixes #5018.