@ledvinap 2018-02-22
@etracer65 : Could be easily implemented as bitmask of enabled timers, but it's fine this way ...