@etracer65 2018-02-22
@ledvinap Yes, I agree that doing it by bitmask would probably be a little cleaner.  I went with this method to try and keep it aligned with `timerVal[[]` and hopefully keep it easier for the next person to see what's going on.  The ledstrip code is pretty convoluted...