@etracer65 2018-02-22
But changing it to use bitmask would save a few bytes of RAM as well...  I have some other PR's coming for feature enhancements to ledstrip.  Maybe I'll change to the bitmask approach in one of those.