@etracer65 2018-02-22
Another possibility to look at is around the actual writes to the LEDs.  `ws2811UpdateStrip()` which does the DMA writes gets called every  time the task runes (100hz) but *most* of the time none of the LEDs have actually changed.  I don't really know much about how DMA operates in this MCU context and whether it matters enough to avoid the DMA writes  when not needed.

Not sure if the performance value would be enough to offset the complication of keeping a "previous" state for comparison.