@etracer65 2018-02-22
I went ahead and did the bitmask changes separately so that it wouldn't be tied to an enhancement that needs to wait for 3.4.

https://github.com/betaflight/betaflight/pull/5270